requirements.txt
PySide6>=6.6.0
openai>=1.40.0
anthropic>=0.34.0
python-dotenv>=1.0.0
